restructure sections
