+++ components/Home/OurOffer.tsx
@@ -7,7 +7,7 @@ import {
   useColorMode,
 } from "@chakra-ui/react";
 
-import { Card } from "../components/Card";
+import { Card } from "../Card";
 
 const offer = [
   {
