+++ components/header/Header.tsx
@@ -1,7 +1,7 @@
 import { Box, Container, Flex } from "@chakra-ui/react";
 import Link from "next/link";
 
-import Logo from "../../components/Logo";
+import Logo from "./Logo";
 import NavigationWrapper from "./navigation/NavigationWrapper";
 
 const Header = () => (
